requirements/dev.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --extra=dev --output-file=requirements/dev.txt setup.cfg
#
asgiref==3.5.0
    # via django
astroid==2.9.3
    # via pylint
asttokens==2.0.5
    # via stack-data
backcall==0.2.0
    # via ipython
black==22.1.0
    # via
    #   ipython
    #   lti-producer (setup.cfg)
click==8.0.4
    # via
    #   black
    #   pip-tools
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
django==3.2.12
    # via
    #   django-debug-toolbar
    #   django-lti-toolbox
    #   lti-producer (setup.cfg)
django-debug-toolbar==3.2.4
    # via lti-producer (setup.cfg)
django-lti-toolbox==1.0.1
    # via lti-producer (setup.cfg)
executing==0.8.2
    # via stack-data
huey==2.4.3
    # via lti-producer (setup.cfg)
ipdb==0.13.9
    # via lti-producer (setup.cfg)
ipython==8.0.1
    # via ipdb
isort==5.10.1
    # via pylint
jedi==0.18.1
    # via ipython
lazy-object-proxy==1.7.1
    # via astroid
matplotlib-inline==0.1.3
    # via ipython
mccabe==0.6.1
    # via pylint
mypy==0.931
    # via lti-producer (setup.cfg)
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
oauthlib==3.2.0
    # via django-lti-toolbox
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pep517==0.12.0
    # via pip-tools
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pip-tools==6.5.1
    # via lti-producer (setup.cfg)
platformdirs==2.5.1
    # via
    #   black
    #   pylint
prompt-toolkit==3.0.28
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pygments==2.11.2
    # via ipython
pylint==2.12.2
    # via
    #   lti-producer (setup.cfg)
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.2
    # via lti-producer (setup.cfg)
pylint-plugin-utils==0.7
    # via pylint-django
pytz==2021.3
    # via django
six==1.16.0
    # via asttokens
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.2.0
    # via ipython
toml==0.10.2
    # via
    #   ipdb
    #   pylint
tomli==2.0.1
    # via
    #   black
    #   mypy
    #   pep517
traitlets==5.1.1
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.1.1
    # via
    #   astroid
    #   black
    #   mypy
    #   pylint
wcwidth==0.2.5
    # via prompt-toolkit
wheel==0.37.1
    # via pip-tools
wrapt==1.13.3
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
